+++ monero-harness/tests/monerod.rs
@@ -14,7 +14,7 @@ async fn init_miner_and_mine_to_miner_address() {
     let tc = Cli::default();
     let (monero, _monerod_container) = Monero::new(&tc, vec![]).await.unwrap();
 
-    monero.init(vec![]).await.unwrap();
+    monero.init_and_start_miner().await.unwrap();
 
     let monerod = monero.monerod();
     let miner_wallet = monero.wallet("miner").unwrap();