+++ swap/tests/alice_punishes_after_restart_punish_timelock_expired.rs
@@ -13,6 +13,7 @@ use swap::protocol::{alice, bob};
 async fn alice_punishes_after_restart_if_punish_timelock_expired() {
     harness::setup_test(FastPunishConfig, |mut ctx| async move {
         let (bob_swap, bob_join_handle) = ctx.bob_swap().await;
+        let bob_swap_id = bob_swap.swap_id;
         let bob_swap = tokio::spawn(bob::run_until(bob_swap, is_btc_locked));
 
         let alice_swap = ctx.alice_next_swap().await;
@@ -33,11 +34,7 @@ async fn alice_punishes_after_restart_if_punish_timelock_expired() {
                 .wait_until_confirmed_with(state3.punish_timelock)
                 .await?;
         } else {
-            panic!(
-                "\
-            Alice in unexpected state {}",
-                alice_state
-            );
+            panic!("Alice in unexpected state {}", alice_state);
         }
 
         ctx.restart_alice().await;
@@ -49,7 +46,9 @@ async fn alice_punishes_after_restart_if_punish_timelock_expired() {
 
         // Restart Bob after Alice punished to ensure Bob transitions to
         // punished and does not run indefinitely
-        let (bob_swap, _) = ctx.stop_and_resume_bob_from_db(bob_join_handle).await;
+        let (bob_swap, _) = ctx
+            .stop_and_resume_bob_from_db(bob_join_handle, bob_swap_id)
+            .await;
         assert!(matches!(bob_swap.state, BobState::BtcLocked { .. }));
 
         let bob_state = bob::run(bob_swap).await?;